+++ packages/core/src/aliases.ts
@@ -1,132 +1,624 @@
 export default {
-  "chatgpt-4o": {
-    "openai": "chatgpt-4o-latest",
-    "openrouter": "openai/chatgpt-4o-latest"
-  },
   "claude-3-5-haiku": {
     "anthropic": "claude-3-5-haiku-20241022",
-    "openrouter": "anthropic/claude-3.5-haiku",
-    "bedrock": "anthropic.claude-3-5-haiku-20241022-v1:0"
+    "openrouter": "anthropic/claude-3.5-haiku"
   },
   "claude-3-5-sonnet": {
-    "anthropic": "claude-3-5-sonnet-20240620",
-    "openrouter": "anthropic/claude-3.5-sonnet",
-    "bedrock": "anthropic.claude-3-5-sonnet-20240620-v1:0"
-  },
-  "claude-sonnet-4-5": {
-    "anthropic": "claude-sonnet-4-5-20250929"
-  },
-  "claude-sonnet-4": {
-    "anthropic": "claude-sonnet-4-20250514"
+    "anthropic": "claude-3-5-sonnet-20241022",
+    "openrouter": "anthropic/claude-3.5-sonnet"
   },
   "claude-3-7-sonnet": {
     "anthropic": "claude-3-7-sonnet-20250219",
+    "openrouter": "anthropic/claude-3.7-sonnet"
+  },
+  "claude-haiku-3": {
+    "anthropic": "claude-3-haiku-20240307"
+  },
+  "claude-haiku-3.5": {
+    "openrouter": "anthropic/claude-3.5-haiku",
+    "anthropic": "claude-3-5-haiku-20241022"
+  },
+  "claude-haiku-3.5-latest": {
+    "anthropic": "claude-3-5-haiku-latest"
+  },
+  "claude-haiku-4.5": {
+    "openrouter": "anthropic/claude-haiku-4.5",
+    "anthropic": "claude-haiku-4-5-20251001"
+  },
+  "claude-haiku-4.5-latest": {
+    "anthropic": "claude-haiku-4-5"
+  },
+  "claude-opus-3": {
+    "anthropic": "claude-3-opus-20240229"
+  },
+  "claude-opus-4": {
+    "openrouter": "anthropic/claude-opus-4",
+    "anthropic": "claude-opus-4-20250514"
+  },
+  "claude-opus-4-latest": {
+    "anthropic": "claude-opus-4-0"
+  },
+  "claude-opus-4.1": {
+    "openrouter": "anthropic/claude-opus-4.1",
+    "anthropic": "claude-opus-4-1-20250805"
+  },
+  "claude-opus-4.1-latest": {
+    "anthropic": "claude-opus-4-1"
+  },
+  "claude-opus-4.5": {
+    "openrouter": "anthropic/claude-opus-4.5",
+    "anthropic": "claude-opus-4-5-20251101"
+  },
+  "claude-opus-4.5-latest": {
+    "anthropic": "claude-opus-4-5"
+  },
+  "claude-sonnet-3": {
+    "anthropic": "claude-3-sonnet-20240229"
+  },
+  "claude-sonnet-3.5": {
+    "anthropic": "claude-3-5-sonnet-20240620"
+  },
+  "claude-sonnet-3.7": {
     "openrouter": "anthropic/claude-3.7-sonnet",
-    "bedrock": "us.anthropic.claude-3-7-sonnet-20250219-v1:0"
+    "anthropic": "claude-3-7-sonnet-20250219"
   },
-  "claude-3-haiku": {
-    "anthropic": "claude-3-haiku-20240307",
-    "openrouter": "anthropic/claude-3-haiku",
-    "bedrock": "anthropic.claude-3-haiku-20240307-v1:0:200k"
+  "claude-sonnet-3.7-latest": {
+    "anthropic": "claude-3-7-sonnet-latest"
   },
-  "claude-3-opus": {
-    "anthropic": "claude-3-opus-20240229",
-    "openrouter": "anthropic/claude-3-opus",
-    "bedrock": "anthropic.claude-3-opus-20240229-v1:0:200k"
+  "claude-sonnet-4": {
+    "openrouter": "anthropic/claude-sonnet-4",
+    "anthropic": "claude-sonnet-4-20250514"
   },
-  "claude-3-sonnet": {
-    "bedrock": "anthropic.claude-3-sonnet-20240229-v1:0"
+  "claude-sonnet-4-latest": {
+    "anthropic": "claude-sonnet-4-0"
+  },
+  "claude-sonnet-4.5": {
+    "openrouter": "anthropic/claude-sonnet-4.5",
+    "anthropic": "claude-sonnet-4-5-20250929"
+  },
+  "claude-sonnet-4.5-latest": {
+    "anthropic": "claude-sonnet-4-5"
+  },
+  "codestral-2508": {
+    "openrouter": "mistralai/codestral-2508"
+  },
+  "codex-mini": {
+    "openai": "codex-mini-latest"
+  },
+  "deephermes-3-llama-3-8b-preview": {
+    "openrouter": "nousresearch/deephermes-3-llama-3-8b-preview"
   },
   "deepseek-chat": {
     "deepseek": "deepseek-chat",
     "openrouter": "deepseek/deepseek-chat"
   },
-  "gemini-flash": {
-    "gemini": "gemini-flash-latest",
-    "vertexai": "gemini-flash-latest",
-    "openrouter": "google/gemini-flash-latest"
+  "deepseek-r1-0528-qwen3-8b-free": {
+    "openrouter": "deepseek/deepseek-r1-0528-qwen3-8b:free"
   },
-  "gemini-pro": {
-    "gemini": "gemini-1.5-pro-001",
-    "vertexai": "gemini-1.5-pro-001",
-    "openrouter": "google/gemini-1.5-pro-001"
+  "deepseek-r1-distill-llama-70b": {
+    "openrouter": "deepseek/deepseek-r1-distill-llama-70b"
+  },
+  "deepseek-r1-distill-qwen-14b": {
+    "openrouter": "deepseek/deepseek-r1-distill-qwen-14b"
+  },
+  "deepseek-r1t2-chimera-free": {
+    "openrouter": "tngtech/deepseek-r1t2-chimera:free"
+  },
+  "deepseek-reasoner": {
+    "deepseek": "deepseek-reasoner",
+    "openrouter": "deepseek/deepseek-reasoner"
+  },
+  "deepseek-v3-0324": {
+    "openrouter": "deepseek/deepseek-chat-v3-0324"
+  },
+  "deepseek-v3-base-free": {
+    "openrouter": "deepseek/deepseek-v3-base:free"
+  },
+  "deepseek-v3.1": {
+    "openrouter": "deepseek/deepseek-chat-v3.1"
+  },
+  "deepseek-v3.1-terminus": {
+    "openrouter": "deepseek/deepseek-v3.1-terminus"
+  },
+  "deepseek-v3.1-terminus-exacto": {
+    "openrouter": "deepseek/deepseek-v3.1-terminus:exacto"
+  },
+  "deepseek-v3.2": {
+    "openrouter": "deepseek/deepseek-v3.2"
+  },
+  "deepseek-v3.2-speciale": {
+    "openrouter": "deepseek/deepseek-v3.2-speciale"
+  },
+  "devstral-2-2512": {
+    "openrouter": "mistralai/devstral-2512"
+  },
+  "devstral-2-2512-free": {
+    "openrouter": "mistralai/devstral-2512:free"
+  },
+  "devstral-medium": {
+    "openrouter": "mistralai/devstral-medium-2507"
+  },
+  "devstral-small": {
+    "openrouter": "mistralai/devstral-small-2505"
+  },
+  "devstral-small-1.1": {
+    "openrouter": "mistralai/devstral-small-2507"
+  },
+  "devstral-small-2505-free": {
+    "openrouter": "mistralai/devstral-small-2505:free"
+  },
+  "dolphin3.0-mistral-24b": {
+    "openrouter": "cognitivecomputations/dolphin3.0-mistral-24b"
+  },
+  "dolphin3.0-r1-mistral-24b": {
+    "openrouter": "cognitivecomputations/dolphin3.0-r1-mistral-24b"
   },
   "gemini-1.5-flash": {
-    "gemini": "gemini-1.5-flash-001",
-    "vertexai": "gemini-1.5-flash-001",
-    "openrouter": "google/gemini-1.5-flash-001"
+    "gemini": "gemini-1.5-flash"
+  },
+  "gemini-1.5-flash-8b": {
+    "gemini": "gemini-1.5-flash-8b"
   },
   "gemini-1.5-pro": {
-    "gemini": "gemini-1.5-pro-001",
-    "vertexai": "gemini-1.5-pro-001",
-    "openrouter": "google/gemini-1.5-pro-001"
+    "gemini": "gemini-1.5-pro",
+    "openrouter": "google/gemini-pro-1.5"
   },
   "gemini-2.0-flash": {
     "gemini": "gemini-2.0-flash",
-    "vertexai": "gemini-2.0-flash"
+    "openrouter": "google/gemini-2.0-flash-001"
   },
-  "gemini-2.0-flash-001": {
-    "gemini": "gemini-2.0-flash-001",
-    "openrouter": "google/gemini-2.0-flash-001",
-    "vertexai": "gemini-2.0-flash-001"
+  "gemini-2.0-flash-experimental-free": {
+    "openrouter": "google/gemini-2.0-flash-exp:free"
+  },
+  "gemini-2.0-flash-lite": {
+    "gemini": "gemini-2.0-flash-lite"
+  },
+  "gemini-2.5-flash": {
+    "gemini": "gemini-2.5-flash",
+    "openrouter": "google/gemini-2.5-flash"
+  },
+  "gemini-2.5-flash-image": {
+    "gemini": "gemini-2.5-flash-image"
+  },
+  "gemini-2.5-flash-image-preview": {
+    "gemini": "gemini-2.5-flash-image-preview"
+  },
+  "gemini-2.5-flash-lite": {
+    "gemini": "gemini-2.5-flash-lite",
+    "openrouter": "google/gemini-2.5-flash-lite"
+  },
+  "gemini-2.5-flash-lite-preview-06-17": {
+    "gemini": "gemini-2.5-flash-lite-preview-06-17"
+  },
+  "gemini-2.5-flash-lite-preview-09-25": {
+    "gemini": "gemini-2.5-flash-lite-preview-09-2025",
+    "openrouter": "google/gemini-2.5-flash-lite-preview-09-2025"
+  },
+  "gemini-2.5-flash-preview-04-17": {
+    "gemini": "gemini-2.5-flash-preview-04-17"
+  },
+  "gemini-2.5-flash-preview-05-20": {
+    "gemini": "gemini-2.5-flash-preview-05-20"
+  },
+  "gemini-2.5-flash-preview-09-25": {
+    "gemini": "gemini-2.5-flash-preview-09-2025",
+    "openrouter": "google/gemini-2.5-flash-preview-09-2025"
+  },
+  "gemini-2.5-flash-preview-tts": {
+    "gemini": "gemini-2.5-flash-preview-tts"
+  },
+  "gemini-2.5-pro": {
+    "gemini": "gemini-2.5-pro",
+    "openrouter": "google/gemini-2.5-pro"
+  },
+  "gemini-2.5-pro-preview-05-06": {
+    "gemini": "gemini-2.5-pro-preview-05-06",
+    "openrouter": "google/gemini-2.5-pro-preview-05-06"
+  },
+  "gemini-2.5-pro-preview-06-05": {
+    "gemini": "gemini-2.5-pro-preview-06-05",
+    "openrouter": "google/gemini-2.5-pro-preview-06-05"
+  },
+  "gemini-2.5-pro-preview-tts": {
+    "gemini": "gemini-2.5-pro-preview-tts"
+  },
+  "gemini-3-flash-preview": {
+    "gemini": "gemini-3-flash-preview",
+    "openrouter": "google/gemini-3-flash-preview"
+  },
+  "gemini-3-pro-preview": {
+    "gemini": "gemini-3-pro-preview",
+    "openrouter": "google/gemini-3-pro-preview"
+  },
+  "gemini-embedding-001": {
+    "gemini": "gemini-embedding-001"
+  },
+  "gemini-flash-latest": {
+    "gemini": "gemini-flash-latest"
+  },
+  "gemini-flash-lite-latest": {
+    "gemini": "gemini-flash-lite-latest"
+  },
+  "gemini-live-2.5-flash": {
+    "gemini": "gemini-live-2.5-flash"
+  },
+  "gemini-live-2.5-flash-preview-native-audio": {
+    "gemini": "gemini-live-2.5-flash-preview-native-audio"
+  },
+  "gemma-2-9b-free": {
+    "openrouter": "google/gemma-2-9b-it:free"
+  },
+  "gemma-3-12b-it": {
+    "openrouter": "google/gemma-3-12b-it"
+  },
+  "gemma-3-27b-it": {
+    "openrouter": "google/gemma-3-27b-it"
+  },
+  "gemma-3n-4b-free": {
+    "openrouter": "google/gemma-3n-e4b-it:free"
+  },
+  "gemma-3n-e4b-it": {
+    "openrouter": "google/gemma-3n-e4b-it"
+  },
+  "glm-4.5": {
+    "openrouter": "z-ai/glm-4.5"
+  },
+  "glm-4.5-air": {
+    "openrouter": "z-ai/glm-4.5-air"
+  },
+  "glm-4.5-air-free": {
+    "openrouter": "z-ai/glm-4.5-air:free"
+  },
+  "glm-4.5v": {
+    "openrouter": "z-ai/glm-4.5v"
+  },
+  "glm-4.6": {
+    "openrouter": "z-ai/glm-4.6"
+  },
+  "glm-4.6-exacto": {
+    "openrouter": "z-ai/glm-4.6:exacto"
+  },
+  "glm-4.7": {
+    "openrouter": "z-ai/glm-4.7"
+  },
+  "glm-z1-32b-free": {
+    "openrouter": "thudm/glm-z1-32b:free"
   },
   "gpt-3.5-turbo": {
-    "openai": "gpt-3.5-turbo",
-    "openrouter": "openai/gpt-3.5-turbo"
+    "openai": "gpt-3.5-turbo"
   },
   "gpt-4": {
-    "openai": "gpt-4",
-    "openrouter": "openai/gpt-4"
+    "openai": "gpt-4"
   },
   "gpt-4-turbo": {
-    "openai": "gpt-4-turbo",
-    "openrouter": "openai/gpt-4-turbo"
+    "openai": "gpt-4-turbo"
+  },
+  "gpt-4.1": {
+    "openai": "gpt-4.1",
+    "openrouter": "openai/gpt-4.1"
+  },
+  "gpt-4.1-mini": {
+    "openai": "gpt-4.1-mini",
+    "openrouter": "openai/gpt-4.1-mini"
+  },
+  "gpt-4.1-nano": {
+    "openai": "gpt-4.1-nano"
   },
   "gpt-4o": {
     "openai": "gpt-4o",
     "openrouter": "openai/gpt-4o"
   },
+  "gpt-4o-2024-05-13": {
+    "openai": "gpt-4o-2024-05-13"
+  },
+  "gpt-4o-2024-08-06": {
+    "openai": "gpt-4o-2024-08-06"
+  },
+  "gpt-4o-2024-11-20": {
+    "openai": "gpt-4o-2024-11-20"
+  },
   "gpt-4o-mini": {
     "openai": "gpt-4o-mini",
     "openrouter": "openai/gpt-4o-mini"
   },
-  "llama-3-1-405b": {
-    "openrouter": "meta-llama/llama-3.1-405b"
+  "gpt-5": {
+    "openai": "gpt-5",
+    "openrouter": "openai/gpt-5"
   },
-  "llama-3-1-405b-instruct": {
-    "openrouter": "meta-llama/llama-3.1-405b-instruct"
+  "gpt-5-chat-latest": {
+    "openai": "gpt-5-chat-latest",
+    "openrouter": "openai/gpt-5-chat"
   },
-  "llama-3-1-70b": {
-    "openrouter": "meta-llama/llama-3.1-70b"
+  "gpt-5-codex": {
+    "openai": "gpt-5-codex",
+    "openrouter": "openai/gpt-5-codex"
   },
-  "llama-3-1-70b-instruct": {
-    "openrouter": "meta-llama/llama-3.1-70b-instruct"
+  "gpt-5-image": {
+    "openrouter": "openai/gpt-5-image"
   },
-  "llama-3-1-8b": {
-    "openrouter": "meta-llama/llama-3.1-8b"
+  "gpt-5-mini": {
+    "openai": "gpt-5-mini",
+    "openrouter": "openai/gpt-5-mini"
   },
-  "llama-3-1-8b-instruct": {
-    "openrouter": "meta-llama/llama-3.1-8b-instruct"
+  "gpt-5-nano": {
+    "openai": "gpt-5-nano",
+    "openrouter": "openai/gpt-5-nano"
   },
-  "llama-3-2-1b-instruct": {
-    "openrouter": "meta-llama/llama-3.2-1b-instruct"
+  "gpt-5-pro": {
+    "openai": "gpt-5-pro",
+    "openrouter": "openai/gpt-5-pro"
   },
-  "llama-3-2-3b-instruct": {
-    "openrouter": "meta-llama/llama-3.2-3b-instruct"
+  "gpt-5.1": {
+    "openai": "gpt-5.1",
+    "openrouter": "openai/gpt-5.1"
   },
-  "llama-3-3-70b-instruct": {
-    "openrouter": "meta-llama/llama-3.3-70b-instruct"
+  "gpt-5.1-chat": {
+    "openai": "gpt-5.1-chat-latest",
+    "openrouter": "openai/gpt-5.1-chat"
   },
-  "mistral-large": {
-    "mistral": "mistral-large-latest",
-    "openrouter": "mistralai/mistral-large"
+  "gpt-5.1-codex": {
+    "openai": "gpt-5.1-codex",
+    "openrouter": "openai/gpt-5.1-codex"
   },
-  "mistral-medium": {
-    "mistral": "mistral-medium-latest",
-    "openrouter": "mistralai/mistral-medium"
+  "gpt-5.1-codex-max": {
+    "openai": "gpt-5.1-codex-max"
   },
-  "mistral-small": {
-    "mistral": "mistral-small-latest",
-    "openrouter": "mistralai/mistral-small"
+  "gpt-5.1-codex-mini": {
+    "openai": "gpt-5.1-codex-mini",
+    "openrouter": "openai/gpt-5.1-codex-mini"
+  },
+  "gpt-5.2": {
+    "openai": "gpt-5.2",
+    "openrouter": "openai/gpt-5.2"
+  },
+  "gpt-5.2-chat": {
+    "openai": "gpt-5.2-chat-latest",
+    "openrouter": "openai/gpt-5.2-chat-latest"
+  },
+  "gpt-5.2-pro": {
+    "openai": "gpt-5.2-pro",
+    "openrouter": "openai/gpt-5.2-pro"
+  },
+  "gpt-oss-120b": {
+    "gemini": "openai/gpt-oss-120b-maas",
+    "openrouter": "openai/gpt-oss-120b"
+  },
+  "gpt-oss-120b-exacto": {
+    "openrouter": "openai/gpt-oss-120b:exacto"
+  },
+  "gpt-oss-20b": {
+    "gemini": "openai/gpt-oss-20b-maas",
+    "openrouter": "openai/gpt-oss-20b"
+  },
+  "gpt-oss-safeguard-20b": {
+    "openrouter": "openai/gpt-oss-safeguard-20b"
+  },
+  "grok-3": {
+    "openrouter": "x-ai/grok-3"
+  },
+  "grok-3-beta": {
+    "openrouter": "x-ai/grok-3-beta"
+  },
+  "grok-3-mini": {
+    "openrouter": "x-ai/grok-3-mini"
+  },
+  "grok-3-mini-beta": {
+    "openrouter": "x-ai/grok-3-mini-beta"
+  },
+  "grok-4": {
+    "openrouter": "x-ai/grok-4"
+  },
+  "grok-4-fast": {
+    "openrouter": "x-ai/grok-4-fast"
+  },
+  "grok-4.1-fast": {
+    "openrouter": "x-ai/grok-4.1-fast"
+  },
+  "grok-code-fast-1": {
+    "openrouter": "x-ai/grok-code-fast-1"
+  },
+  "hermes-4-405b": {
+    "openrouter": "nousresearch/hermes-4-405b"
+  },
+  "hermes-4-70b": {
+    "openrouter": "nousresearch/hermes-4-70b"
+  },
+  "kat-coder-pro-free": {
+    "openrouter": "kwaipilot/kat-coder-pro:free"
+  },
+  "kimi-dev-72b-free": {
+    "openrouter": "moonshotai/kimi-dev-72b:free"
+  },
+  "kimi-k2": {
+    "openrouter": "moonshotai/kimi-k2"
+  },
+  "kimi-k2-free": {
+    "openrouter": "moonshotai/kimi-k2:free"
+  },
+  "kimi-k2-instruct-0905": {
+    "openrouter": "moonshotai/kimi-k2-0905"
+  },
+  "kimi-k2-instruct-0905-exacto": {
+    "openrouter": "moonshotai/kimi-k2-0905:exacto"
+  },
+  "kimi-k2-thinking": {
+    "openrouter": "moonshotai/kimi-k2-thinking"
+  },
+  "llama-3.2-11b-vision-instruct": {
+    "openrouter": "meta-llama/llama-3.2-11b-vision-instruct"
+  },
+  "llama-3.3-70b-instruct-free": {
+    "openrouter": "meta-llama/llama-3.3-70b-instruct:free"
+  },
+  "llama-4-scout-free": {
+    "openrouter": "meta-llama/llama-4-scout:free"
+  },
+  "mai-ds-r1-free": {
+    "openrouter": "microsoft/mai-ds-r1:free"
+  },
+  "minimax-01": {
+    "openrouter": "minimax/minimax-01"
+  },
+  "minimax-m1": {
+    "openrouter": "minimax/minimax-m1"
+  },
+  "minimax-m2": {
+    "openrouter": "minimax/minimax-m2"
+  },
+  "minimax-m2.1": {
+    "openrouter": "minimax/minimax-m2.1"
+  },
+  "mistral-7b-instruct-free": {
+    "openrouter": "mistralai/mistral-7b-instruct:free"
+  },
+  "mistral-medium-3": {
+    "openrouter": "mistralai/mistral-medium-3"
+  },
+  "mistral-medium-3.1": {
+    "openrouter": "mistralai/mistral-medium-3.1"
+  },
+  "mistral-nemo-free": {
+    "openrouter": "mistralai/mistral-nemo:free"
+  },
+  "mistral-small-3.1-24b-instruct": {
+    "openrouter": "mistralai/mistral-small-3.1-24b-instruct"
+  },
+  "mistral-small-3.2-24b-free": {
+    "openrouter": "mistralai/mistral-small-3.2-24b-instruct:free"
+  },
+  "mistral-small-3.2-24b-instruct": {
+    "openrouter": "mistralai/mistral-small-3.2-24b-instruct"
+  },
+  "nvidia-nemotron-nano-9b": {
+    "openrouter": "nvidia/nemotron-nano-9b-v2"
+  },
+  "o1": {
+    "openai": "o1",
+    "openrouter": "openai/o1"
+  },
+  "o1-mini": {
+    "openai": "o1-mini"
+  },
+  "o1-preview": {
+    "openai": "o1-preview"
+  },
+  "o1-pro": {
+    "openai": "o1-pro"
+  },
+  "o3-deep-research": {
+    "openai": "o3-deep-research"
+  },
+  "o3-mini": {
+    "openai": "o3-mini",
+    "openrouter": "openai/o3-mini"
+  },
+  "o3-pro": {
+    "openai": "o3-pro"
+  },
+  "o4-mini": {
+    "openai": "o4-mini",
+    "openrouter": "openai/o4-mini"
+  },
+  "o4-mini-deep-research": {
+    "openai": "o4-mini-deep-research"
+  },
+  "qwen2.5-coder-32b-instruct": {
+    "openrouter": "qwen/qwen-2.5-coder-32b-instruct"
+  },
+  "qwen2.5-vl-32b-instruct-free": {
+    "openrouter": "qwen/qwen2.5-vl-32b-instruct:free"
+  },
+  "qwen2.5-vl-72b-instruct": {
+    "openrouter": "qwen/qwen2.5-vl-72b-instruct"
+  },
+  "qwen2.5-vl-72b-instruct-free": {
+    "openrouter": "qwen/qwen2.5-vl-72b-instruct:free"
+  },
+  "qwen3-14b-free": {
+    "openrouter": "qwen/qwen3-14b:free"
+  },
+  "qwen3-235b-a22b-free": {
+    "openrouter": "qwen/qwen3-235b-a22b:free"
+  },
+  "qwen3-235b-a22b-instruct-2507": {
+    "openrouter": "qwen/qwen3-235b-a22b-07-25"
+  },
+  "qwen3-235b-a22b-instruct-2507-free": {
+    "openrouter": "qwen/qwen3-235b-a22b-07-25:free"
+  },
+  "qwen3-235b-a22b-thinking-2507": {
+    "openrouter": "qwen/qwen3-235b-a22b-thinking-2507"
+  },
+  "qwen3-30b-a3b-free": {
+    "openrouter": "qwen/qwen3-30b-a3b:free"
+  },
+  "qwen3-30b-a3b-instruct-2507": {
+    "openrouter": "qwen/qwen3-30b-a3b-instruct-2507"
+  },
+  "qwen3-30b-a3b-thinking-2507": {
+    "openrouter": "qwen/qwen3-30b-a3b-thinking-2507"
+  },
+  "qwen3-32b-free": {
+    "openrouter": "qwen/qwen3-32b:free"
+  },
+  "qwen3-8b-free": {
+    "openrouter": "qwen/qwen3-8b:free"
+  },
+  "qwen3-coder": {
+    "openrouter": "qwen/qwen3-coder"
+  },
+  "qwen3-coder-30b-a3b-instruct": {
+    "openrouter": "qwen/qwen3-coder-30b-a3b-instruct"
+  },
+  "qwen3-coder-480b-a35b-instruct-free": {
+    "openrouter": "qwen/qwen3-coder:free"
+  },
+  "qwen3-coder-exacto": {
+    "openrouter": "qwen/qwen3-coder:exacto"
+  },
+  "qwen3-coder-flash": {
+    "openrouter": "qwen/qwen3-coder-flash"
+  },
+  "qwen3-max": {
+    "openrouter": "qwen/qwen3-max"
+  },
+  "qwen3-next-80b-a3b-instruct": {
+    "openrouter": "qwen/qwen3-next-80b-a3b-instruct"
+  },
+  "qwen3-next-80b-a3b-thinking": {
+    "openrouter": "qwen/qwen3-next-80b-a3b-thinking"
+  },
+  "qwerky-72b": {
+    "openrouter": "featherless/qwerky-72b"
+  },
+  "qwq-32b-free": {
+    "openrouter": "qwen/qwq-32b:free"
+  },
+  "r1-0528-free": {
+    "openrouter": "deepseek/deepseek-r1-0528:free"
+  },
+  "r1-free": {
+    "openrouter": "deepseek/deepseek-r1:free"
+  },
+  "reka-flash-3": {
+    "openrouter": "rekaai/reka-flash-3"
+  },
+  "sarvam-m-free": {
+    "openrouter": "sarvamai/sarvam-m:free"
+  },
+  "sherlock-dash-alpha": {
+    "openrouter": "openrouter/sherlock-dash-alpha"
+  },
+  "sherlock-think-alpha": {
+    "openrouter": "openrouter/sherlock-think-alpha"
+  },
+  "text-embedding-3-large": {
+    "openai": "text-embedding-3-large"
+  },
+  "text-embedding-3-small": {
+    "openai": "text-embedding-3-small"
+  },
+  "text-embedding-ada-002": {
+    "openai": "text-embedding-ada-002"
   }
 } as const;